requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
multidict==5.1.0
pipenv==2020.11.15
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.2.2
virtualenv-clone==0.5.4
yarl==1.6.3
celery~=5.0.5